Bracu Shout Out
Our project “Bracu Shout-Out” has been created to make university life better in a orderly way for both bracu students and faculties. This is an well-controlled database for keeping the track of classes, sections and time schedules. Students and faculties can communicate with each other through this platform for any announcement or queries regarding the courses.